pyproject.toml
[tool.poetry]
name = "code_coil"
version = "0.1.0"
description = "A curated Python Q&A dataset"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "code_coil" }]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1" # Needed for the one-time migration
pandas = "^2.0.3"   # For data analysis

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
ruff = "*"
black = "*"
mypy = "*"
beartype = "*"
types-PyYAML = "*"
tomli-w = "*"


[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B", "W", "D"]

[tool.mypy]
strict = true
packages = "code_coil"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
